ViewData added and MsSQL configuration added to the project

+++ Repositories/Extensions/ProductRepositoryExtension.cs
@@ -21,7 +21,7 @@ namespace Repositories.Extensions
                 .Contains(searchTerm.ToLower()));
         }
 
-        public static IQueryable<Product> FilteredByPrice(this IQueryable<Product> products, int? minPrice, int? maxPrice, bool isValidPrice)
+        public static IQueryable<Product> FilteredByPrice(this IQueryable<Product> products, decimal? minPrice, decimal? maxPrice, bool isValidPrice)
         {
             if (isValidPrice)
                 return products.Where(prd => prd.Price >= minPrice && prd.Price <= maxPrice);
